+++ x/inflation/keeper/inflation.go
@@ -22,7 +22,6 @@ import (
 	evmostypes "github.com/evmos/evmos/v12/types"
 
 	utils "github.com/evmos/evmos/v12/utils"
-	incentivestypes "github.com/evmos/evmos/v12/x/incentives/types"
 	"github.com/evmos/evmos/v12/x/inflation/types"
 )
 
@@ -83,18 +82,6 @@ func (k Keeper) AllocateExponentialInflation(
 		types.ModuleName,
 		k.feeCollectorName,
 		staking,
-	); err != nil {
-		return nil, nil, nil, err
-	}
-
-	// Allocate usage incentives to incentives module account
-	incentives = sdk.Coins{k.GetProportions(ctx, mintedCoin, distribution.UsageIncentives)}
-
-	if err = k.bankKeeper.SendCoinsFromModuleToModule(
-		ctx,
-		types.ModuleName,
-		incentivestypes.ModuleName,
-		incentives,
 	); err != nil {
 		return nil, nil, nil, err
 	}